+++ react/__tests__/Payment.test.tsx
@@ -4,8 +4,7 @@ import { render } from '@vtex/test-tools/react'
 // @ts-ignore
 import MockRouter from 'react-mock-router'
 
-import SubscriptionDetails from '../components/pages/Details'
-import { SubscriptionStatus } from '../constants'
+import SubscriptionDetails from '../components/Details'
 import { mockRouterParam, generateDetailMock } from '../mocks'
 
 const INVALID_PAYMENT =
@@ -62,7 +61,7 @@ describe('Payment Scenarios', () => {
           mocks: [
             generateDetailMock({
               hasPaymentMethod: false,
-              status: SubscriptionStatus.Paused,
+              status: 'PAUSED',
             }),
           ],
         },